requirements.txt
# Core ML Libraries
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0
tensorflow>=2.13.0
keras>=2.13.1

# API Gateway
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.5.0

# Data Storage & Cache
redis==5.0.0

# Monitoring
prometheus-client==0.19.0

# Configuration
pyyaml==6.0.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dateutil==2.8.2
joblib==1.3.2

# Testing
pytest==7.4.3
requests==2.31.0